dbeaver/pro#1325 move registry to model class
