+++ plugins/org.jkiss.dbeaver.model/META-INF/MANIFEST.MF
@@ -11,11 +11,11 @@ Bundle-Activator: org.jkiss.dbeaver.bundle.ModelActivator
 Export-Package: org.jkiss.dbeaver,
  org.jkiss.dbeaver.model,
  org.jkiss.dbeaver.model.access,
- org.jkiss.dbeaver.model.admin.sessions,
  org.jkiss.dbeaver.model.admin.locks,
+ org.jkiss.dbeaver.model.admin.sessions,
+ org.jkiss.dbeaver.model.app,
  org.jkiss.dbeaver.model.auth,
  org.jkiss.dbeaver.model.auth.impl,
- org.jkiss.dbeaver.model.app,
  org.jkiss.dbeaver.model.connection,
  org.jkiss.dbeaver.model.data,
  org.jkiss.dbeaver.model.data.aggregate,
@@ -78,6 +78,7 @@ Export-Package: org.jkiss.dbeaver,
  org.jkiss.dbeaver.model.virtual,
  org.jkiss.dbeaver.registry.expressions,
  org.jkiss.dbeaver.registry.formatter,
+ org.jkiss.dbeaver.registry.timezone,
  org.jkiss.dbeaver.runtime,
  org.jkiss.dbeaver.runtime.encode,
  org.jkiss.dbeaver.runtime.jobs,
