+++ plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/registry/timezone/TimezoneRegistry.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jkiss.dbeaver.registry;
+package org.jkiss.dbeaver.registry.timezone;
 
 import org.jkiss.code.NotNull;
 import org.jkiss.code.Nullable;
@@ -78,7 +78,7 @@ public class TimezoneRegistry {
 
     @NotNull
     public static String getUserDefaultTimezone() {
-        return userDefaultTimezone;
+        return "".equals(userDefaultTimezone) ? TimeZone.getDefault().getID() : userDefaultTimezone;
     }
 
     @NotNull
